+++ cctools/libstuff/lto.c
@@ -63,7 +63,7 @@ size_t size)
 	 *"llvm bitcode file" could be in an archive or fat file.
 	 */
 
-	if(is_llvm_bitcode_from_memory(addr, size, &arch_flag,
+	if(is_llvm_bitcode_from_memory(addr, (uint32_t)size, &arch_flag,
 				       &ofile->lto) != 0){
 	    ofile->lto_cputype = arch_flag.cputype;
 	    ofile->lto_cpusubtype = arch_flag.cpusubtype;
